requirement.txt
streamlit
rdkit-pypi
scikit-learn
numpy
